+++ large_image/config.py
@@ -11,6 +11,7 @@ ConfigValues = {
     'logger': fallbackLogger,
     'logprint': fallbackLogger,
 
+    # For tiles
     'cache_backend': 'python',  # 'python' or 'memcached'
     # 'python' cache can use 1/(val) of the available memory
     'cache_python_memory_portion': 32,
@@ -18,6 +19,15 @@ ConfigValues = {
     'cache_memcached_url': '127.0.0.1',
     'cache_memcached_username': None,
     'cache_memcached_password': None,
+
+    # Generally, these keys are the form of "cache_<cacheName>_<key>"
+
+    # For tilesources.  These are also limited by available file handles.
+    # 'python' cache can use 1/(val) of the available memory based on a very
+    # rough estimate of the amount of memory used by a tilesource
+    'cache_tilesource_memory_portion': 8,
+    # If >0, this is the maximum number of tilesources that will be cached
+    'cache_tilesource_maximum': 0,
 
     'max_small_image_size': 4096,
 }